General 2
Commented out useless print statements, made it so picking up resource tiles gives 5 resources.
Found unreproducible error, but am 80% sure that it was this: https://www.scienceabc.com/innovation/what-are-bit-flips-and-how-are-spacecraft-protected-from-them.html
It happened twice but there's no reason for the error to have happened.
The error:
Traceback (most recent call last):
  File "c:\Users\Dylan\Documents\GitHub\SGAI-Uly-1\SGAI_MK3\main.py", line 41, in <module>
    action = PF.get_action(GameBoard, x, y)
  File "c:\Users\Dylan\Documents\GitHub\SGAI-Uly-1\SGAI_MK3\PygameFunctions.py", line 273, in get_action
    state = GameBoard.States[gridPos[1]][gridPos[0]]
IndexError: list index out of range

The gridPos variable has had a 100% track record of accuracy until now, so it's weird that this error happened twice